Added lesson 2-2_Local_Deployment.ipynb
