+++ src/steps/mlflow_trainer.py
@@ -1,7 +1,33 @@
+import mlflow
+import numpy as np
+from sklearn.base import ClassifierMixin
+from sklearn.svm import SVC
+from sklearn.tree import DecisionTreeClassifier
 from zenml.integrations.mlflow.mlflow_step_decorator import enable_mlflow
+from zenml.steps import step
 
-from steps.trainer import svc_trainer_mlflow, tree_trainer_with_mlflow
 
-# These are the same steps, but with mlflow enabled
-svc_trainer_mlflow = enable_mlflow(svc_trainer_mlflow)
-tree_trainer_with_mlflow = enable_mlflow(tree_trainer_with_mlflow)
+@enable_mlflow  # setup MLflow
+@step(enable_cache=False)
+def svc_trainer_mlflow(
+    X_train: np.ndarray,
+    y_train: np.ndarray,
+) -> ClassifierMixin:
+    """Train a sklearn SVC classifier and log to MLflow."""
+    mlflow.sklearn.autolog()  # log all model hparams and metrics to MLflow
+    model = SVC(gamma=0.001)
+    model.fit(X_train, y_train)
+    return model
+
+
+@enable_mlflow  # setup MLflow
+@step(enable_cache=False)
+def tree_trainer_with_mlflow(
+    X_train: np.ndarray,
+    y_train: np.ndarray,
+) -> ClassifierMixin:
+    """Train a sklearn decision tree classifier and log to MLflow."""
+    mlflow.sklearn.autolog()  # log all model hparams and metrics to MLflow
+    model = DecisionTreeClassifier()
+    model.fit(X_train, y_train)
+    return model
